+++ examples/sep24/sep24.ts
@@ -1,5 +1,6 @@
 import axios from "axios";
-import { Memo, MemoText } from "stellar-sdk";
+import readline from "readline";
+
 import {
   walletSdk,
   Anchor,
@@ -7,9 +8,58 @@ import {
   Types,
   IssuedAssetId,
 } from "@stellar/typescript-wallet-sdk";
+import { Memo, MemoText } from "stellar-sdk";
 
 const wallet = walletSdk.Wallet.TestNet();
 const stellar = wallet.stellar();
+const anchor = wallet.anchor({ homeDomain: "testanchor.stellar.org" });
+const account = stellar.account();
+
+let kp: SigningKeypair;
+
+const runSep24 = async () => {
+  const asset = new IssuedAssetId(
+    "USDC",
+    "GBBD47IF6LWK7P7MDEVSCWR7DPUWV3NY3DTQEVFL4NAT4AQH3ZLLFLA5",
+  );
+
+  // Create Account
+
+  const createAccount = async () => {
+    console.log("creating account ...");
+    kp = account.createKeypair();
+    console.log(`kp: \n${kp.publicKey}\n${kp.secretKey}`);
+    await axios.get("https://friendbot.stellar.org/?addr=" + kp.publicKey);
+
+    const txBuilder = await stellar.transaction({
+      sourceAddress: kp,
+      baseFee: 1000,
+    });
+    const tx = txBuilder.addAssetSupport(asset).build();
+    kp.sign(tx);
+    const success = await stellar.submitTransaction(tx);
+    if (!success) {
+      throw new Error("adding trustline failed");
+    }
+  };
+
+  await createAccount();
+  await runDeposit(anchor, kp);
+  await runDepositWatcher(anchor);
+
+  while (!depositDone) {
+    await new Promise((resolve) => setTimeout(resolve, 1000));
+  }
+
+  const ans = await askQuestion("Do you want to start withdrawal? (y/n)");
+  if (ans !== "y") {
+    process.exit(0);
+  }
+
+  await runWithdraw(anchor, kp);
+  await runWithdrawWatcher(anchor, kp);
+};
+runSep24();
 
 // Create Deposit
 let authTokenDeposit: string;
@@ -140,4 +190,18 @@ export const runWithdrawWatcher = (anchor, kp) => {
   });
 
   stop = resp.stop;
+};
+
+export const askQuestion = (query) => {
+  const rl = readline.createInterface({
+    input: process.stdin,
+    output: process.stdout,
+  });
+
+  return new Promise((resolve) =>
+    rl.question(query, (ans) => {
+      rl.close();
+      resolve(ans);
+    }),
+  );
 };
